lsコマンド条件分岐修正
