+++ 05.ls/ls2.rb
@@ -1,0 +1,8 @@
+#!/usr/bin/env ruby
+
+# frozen_string_literal: true
+
+# コマンドライン引数使うため
+require 'optparse'
+# lオプションのオーナー名とグループ名出すのに必要
+require 'etc'
